Renamed the files of the components
Renamed all the JS files to comoponent.js and all the main
style files of the components to style.css.
