+++ index.html
@@ -19,7 +19,7 @@
     <statistics-card></statistics-card>
 
     <template id="settings-card">
-        <link rel="stylesheet" href="./components/SettingsCard/SettingsCard.css">
+        <link rel="stylesheet" href="./components/SettingsCard/style.css">
         <link rel="stylesheet" href="./fontawesome/css/all.css">
         <details open class="open">
             <summary>
@@ -141,7 +141,7 @@
     </template>
 
     <template id="packet-visualization">
-        <link rel="stylesheet" href="./components/PacketVisualization/PacketVisualization.css">
+        <link rel="stylesheet" href="./components/PacketVisualization/style.css">
         <link rel="stylesheet" href="./fontawesome/css/all.css">
         <section class="card">
             <div class="card-header">
@@ -192,7 +192,7 @@
     </template>
 
     <template id="packet-track">
-        <link rel="stylesheet" href="./components/PacketTrack/PacketTrack.css">
+        <link rel="stylesheet" href="./components/PacketTrack/style.css">
         <svg id="root" viewBox="0 0 100 900">
             <rect id="sender" class="pkt buffered" y="26" x="16" width="68" height="100"/>
             <g id="timer">
@@ -205,7 +205,7 @@
     </template>
 
     <template id="statistics-card">
-        <link rel="stylesheet" href="./components/StatisticsCard/StatisticsCard.css">
+        <link rel="stylesheet" href="./components/StatisticsCard/style.css">
         <section class="card">
             <h2>Live statistics</h2>
             <div id="stats-content">
